# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Solo Energia Client Portal API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: projects
    description: Client project operations
  - name: documents
    description: Document operations
  - name: notifications
    description: Notification operations
  - name: webhooks
    description: Jestor webhook endpoints
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/request-otp:
    post:
      operationId: requestOtp
      tags: [auth]
      summary: Request a one-time login code via email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestOtpBody"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify the OTP code and create a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpBody"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyOtpResponse"
        "401":
          description: Invalid or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Log out and invalidate session
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List the authenticated client's projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a single project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents for the authenticated user's project
      parameters:
        - name: projectId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"

  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - name: projectId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"

  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  /webhooks/jestor/project:
    post:
      operationId: jestorProjectWebhook
      tags: [webhooks]
      summary: Unified Jestor webhook for project creation and updates
      security:
        - webhookSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JestorWebhookPayload"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResult"
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResult"

  /jestor/sync/{jestorId}:
    get:
      operationId: syncJestorProject
      tags: [projects]
      summary: Manually sync a project from the Jestor API
      parameters:
        - name: jestorId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project synced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found in Jestor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  securitySchemes:
    webhookSecret:
      type: apiKey
      in: header
      name: x-webhook-secret

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Error:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    RequestOtpBody:
      type: object
      properties:
        email:
          type: string
      required:
        - email

    VerifyOtpBody:
      type: object
      properties:
        email:
          type: string
        code:
          type: string
      required:
        - email
        - code

    VerifyOtpResponse:
      type: object
      properties:
        message:
          type: string
        projectId:
          type: integer
      required:
        - message
        - projectId

    AuthUser:
      type: object
      properties:
        projectId:
          type: integer
        clientName:
          type: string
        clientEmail:
          type: string
      required:
        - projectId
        - clientName
        - clientEmail

    Project:
      type: object
      properties:
        id:
          type: integer
        clientName:
          type: string
        clientEmail:
          type: string
        clientPhone:
          type: string
          nullable: true
        systemPower:
          type: number
          description: System power in kWp
        statusStep:
          type: integer
          description: "Current step: 1=Onboarding, 2=Engenharia, 3=Homologação, 4=Logística, 5=Execução, 6=Ativação"
        statusProjeto:
          type: string
          nullable: true
          description: Raw Jestor status_projeto string
        trackingCode:
          type: string
          nullable: true
        trackingCarrier:
          type: string
          nullable: true
        city:
          type: string
        state:
          type: string
        completionPercent:
          type: integer
        estimatedActivation:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
          description: Per-step personalized observations for the client
        estimatedDate:
          type: string
          nullable: true
          description: Estimated completion date for the current phase
        valorProjeto:
          type: number
          nullable: true
        formaDePagamento:
          type: string
          nullable: true
        observacoesGerais:
          type: string
          nullable: true
        dataInicioPrevista:
          type: string
          nullable: true
        dataConclusaoPrevista:
          type: string
          nullable: true
        dataDeFechamento:
          type: string
          nullable: true
        dataDePagamento:
          type: string
          nullable: true
        dataDeCompras:
          type: string
          nullable: true
        dataDeEntregaDoEquipamento:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - clientName
        - clientEmail
        - systemPower
        - statusStep
        - city
        - state
        - completionPercent
        - createdAt

    Document:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [pending_upload, available_download]
        category:
          type: string
          enum: [entrada, intra_projeto]
          description: "entrada = docs the client provides; intra_projeto = docs Solo Energia generates"
        required:
          type: boolean
        description:
          type: string
          nullable: true
        fileUrl:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - projectId
        - name
        - type
        - category
        - required
        - createdAt

    Notification:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        message:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - projectId
        - title
        - message
        - read
        - createdAt

    JestorWebhookPayload:
      type: object
      properties:
        jestor_id:
          type: string
        name:
          type: string
          description: Client name from Jestor
        client_email:
          type: string
        client_phone:
          type: string
          nullable: true
        system_power:
          type: number
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        status_projeto:
          type: string
          nullable: true
        data_inicio_prevista:
          type: string
          nullable: true
        data_conclusao_prevista:
          type: string
          nullable: true
        data_de_fechamento:
          type: string
          nullable: true
        data_de_pagamento:
          type: string
          nullable: true
        data_de_compras:
          type: string
          nullable: true
        data_de_entrega_do_equipamento:
          type: string
          nullable: true
        valor_projeto:
          type: number
          nullable: true
        forma_de_pagamento:
          type: string
          nullable: true
        observacoes_gerais:
          type: string
          nullable: true
        tracking_code:
          type: string
          nullable: true
        tracking_carrier:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - jestor_id

    WebhookResult:
      type: object
      properties:
        message:
          type: string
        project_id:
          type: integer
        created:
          type: boolean
        phase:
          type: integer
          nullable: true
        phase_name:
          type: string
          nullable: true
      required:
        - message
        - project_id
        - created