# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Solo Energia Client Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUser {
  projectId: number;
  clientName: string;
  clientEmail: string;
}