# lib/api-zod/src/generated/types/verifyOtpResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Solo Energia Client Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface VerifyOtpResponse {
  message: string;
  projectId: number;
}